# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Zapp Revamp Backend",
        "description": "Zapp Revamp Backend",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": 
...
        "/user/sms-tac/use": {            
            "post": {
                "parameters": [
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "Give contact number and Tac",
                        "schema": {
                            "$ref": "#/definitions/TacVerificationPayload"
                        }
                    }
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Tac Verification"
                    }
                }
            }
            
        },
...
        "/user/invite-code/use": {            
            "post": {
                "parameters": [
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "In body payload inviteCode,inviteCodeUser",
                        "schema": {
                            "$ref": "#/definitions/UseInviteCode"
                        }
                    }
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "This endpont is for using  Invite code and crediting amount to respective account."
                    }
                }
            }
            
        },
...
    },
    "definitions": {
...
        "TacVerificationPayload": {
            "required": [
                "PhoneNo",
                "tac"
            ],
            "properties": {
                "PhoneNo": {
                    "type": "string"
                },
                "tac": {
                    "type": "string"
                }
            }
        },
...
        "UseInviteCode" : {
            "required": [
                "inviteCode",
                "inviteCodeUser"
            ],
            "properties": {
                "inviteCode": {
                    "type": "string"
                },
                "inviteCodeUser": {
                    "type": "string"
                }
            }
        },
...
    }
}